Add files via upload
In this paper, we consider how to determine the minimum number of leaders with allocation and how to
achieve consensus over directed networks consisting of time-varying nonlinear multi-agents. Firstly, the problem
of finding minimum number of leaders is formulated as a minimum spanning forest problem, i.e., finding the
minimum population of trees in the network. By introducing a toll station connecting with each agent, this problem
is converted to a minimum spanning tree problem. In this way, the minimum number of leaders is determined and
these leaders are found locating at the roots of each tree in the obtained spanning forest. Secondly, we describe a
virtual leader connected with the allocated leaders, which indicates that the number of edges connected the follower
agents with the virtual leader is the least in an arbitrary directed network. This method is different from the existing
consensus problem of redundant leaders or edges that connect the follower with one leader in special networks.
A distributed consensus protocol is revisited for achieving final global consensus of all agents. It is theoretically
shown that such a protocol indeed ensures consensus. Simulation examples in real-life networks are also provided
to show the effectiveness of the proposed methodology. Our works enable studying and extending application of
consensus problems in various complex networks.

The detailed contents can be seen in http://link.springer.com/article/10.1007/s12555-018-0057-2

--------------------------------------------------------------------------------------------------------------------------------------
If your want to reproduce the simulations in this paper, you should run  high_order_nonlinear_consensus.m and switching_consensus.m